+++ mobile/android/android-components/components/feature/media/src/main/java/mozilla/components/feature/media/service/AbstractMediaSessionService.kt
@@ -13,6 +13,7 @@ import android.os.IBinder
 import androidx.annotation.VisibleForTesting
 import mozilla.components.browser.state.state.SessionState
 import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.base.crash.CrashReporting
 import java.lang.ref.WeakReference
 
 /**
@@ -35,6 +36,7 @@ internal class MediaServiceBinder(delegate: MediaSessionDelegate) : Binder() {
  */
 abstract class AbstractMediaSessionService : Service() {
     protected abstract val store: BrowserStore
+    protected abstract val crashReporter: CrashReporting?
 
     @VisibleForTesting
     internal var binder: MediaServiceBinder? = null
@@ -49,6 +51,7 @@ abstract class AbstractMediaSessionService : Service() {
             context = this,
             service = this,
             store = store,
+            crashReporter = crashReporter,
         ).also {
             binder = MediaServiceBinder(it)
         }