+++ mobile/android/android-components/components/feature/media/src/test/java/mozilla/components/feature/media/service/AbstractMediaSessionServiceTest.kt
@@ -9,6 +9,7 @@ import android.content.ComponentName
 import android.content.Intent
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import org.junit.Assert.assertEquals
@@ -93,4 +94,5 @@ class AbstractMediaSessionServiceTest {
 
 class FakeMediaService : AbstractMediaSessionService() {
     public override val store: BrowserStore = mock()
+    public override val crashReporter: CrashReporting = mock()
 }